+++ .vscode/settings.json
@@ -2,5 +2,10 @@
 	"files.associations": {
 		"readline.h": "c",
 		"minishell.h": "c"
-	}
+	},
+	"C_Cpp.files.exclude": {
+    "**/node_modules": true,
+    "**/.git": true,
+    "**/build": true
+  }
 }
