+++ dd.sh
@@ -1,2 +1,1 @@
-cc -g -Wall -Wextra -Werror src/parser.c src/init.c src/quotes/*.c libft/*.c src/util/destroy_split.c -lreadline -o test_bin
-
+cc -g -Wall -Wextra -Werror src/parser/*.c src/init.c src/tokens/*.c libft/*.c src/util/destroy_split.c src/error/*.c -lreadline -o test_bin
