+++ src/tokens/get_paths.c
