+++ test.c
@@ -1,0 +1,1 @@
+(*aux)->token_type != CMD && (*aux)->token_type != ARG
